+++ client/.prettierignore
@@ -1,0 +1,5 @@
+node_modules
+dist
+build
+coverage
+package-lock.json
