+++ client/src/components/agents/AgentConfigurator.tsx
@@ -1,6 +1,7 @@
-import React, { useState } from 'react';
-import { useAgentConfig } from '@/hooks/useAgentConfig';
+import React from 'react'; // useState removed
+// import { useAgentConfig } from '@/hooks/useAgentConfig'; // No longer used directly
 import * as DialogPrimitive from '@radix-ui/react-dialog';
+import { useAgentConfiguratorLogic } from '@/hooks/useAgentConfiguratorLogic'; // Import the new hook
 import {
   AgentType,
   AnyAgentConfig,
@@ -45,6 +46,7 @@ import { mockInitialAgents as importedMockExistingAgents } from '@/data/mocks/mo
 import { ToolSelector } from './tools/ToolSelector'; // Named import
 import AgentList from './AgentList';
 import AgentDropzone from './workflow/AgentDropzone';
+import { createNewAgentConfig } from '@/lib/agent-utils';
 
 // Cast the imported JSON to the Tool array type
 const MOCK_AVAILABLE_TOOLS: Tool[] = mockToolsDataJson as Tool[];
@@ -58,51 +60,6 @@ interface AgentConfiguratorProps {
   isCreatingNew?: boolean;
 }
 
-const createNewAgentConfig = (type: AgentType, existingId?: string, existingName?: string): AnyAgentConfig => {
-  const baseConfig = {
-    id: existingId || crypto.randomUUID(),
-    name: existingName || '',
-  };
-
-  switch (type) {
-    case AgentType.LLM:
-      return {
-        ...baseConfig,
-        type: AgentType.LLM,
-        instruction: '',
-        model: '', // Add model as per shared LlmAgentConfig
-        code_execution: false,
-        planning_enabled: false,
-        tools: [],
-      } as LlmAgentConfig;
-    case AgentType.Sequential:
-      return {
-        ...baseConfig,
-        type: AgentType.Sequential,
-        agents: [], // Will store AnyAgentConfig[]
-      } as SequentialAgentConfig;
-    case AgentType.Parallel:
-      return {
-        ...baseConfig,
-        type: AgentType.Parallel,
-        agents: [], // Will store AnyAgentConfig[] (prop name in type is 'agents')
-      } as ParallelAgentConfig;
-    case AgentType.Loop:
-      // LoopAgentConfig expects an 'agent' property of type AnyAgentConfig.
-      // For creation, it might be null or a placeholder initially.
-      // Or, we find a default agent if applicable, otherwise it's an invalid state until one is selected.
-      return {
-        ...baseConfig,
-        type: AgentType.Loop,
-        // agent: undefined, // Or a default/placeholder if that makes sense
-        // For now, let's assume it can be temporarily invalid until selected
-      } as unknown as LoopAgentConfig; // Cast needed if agent is not set initially
-    default:
-      const _exhaustiveCheck: never = type;
-      throw new Error(`Unknown agent type: ${_exhaustiveCheck}`);
-  }
-};
-
 /**
  * Componente de formulário para edição de agentes.
  * A lógica de estado é delegada ao hook `useAgentConfig`, mantendo a UI simples.
@@ -110,137 +67,47 @@ const createNewAgentConfig = (type: AgentType, existingId?: string, existingName
 const AgentConfigurator: React.FC<AgentConfiguratorProps> = ({
   agentConfig: initialAgentConfig,
   onSave,
-  isSaving = false,
+  isSaving: बाहेरून_येणारे_isSaving = false, // Renamed to avoid conflict with any potential future internal isSaving
   isCreatingNew = false,
 }) => {
   const {
     config,
-    updateConfig,
-    updateField,
-    addTool,
-    removeTool,
-  } = useAgentConfig({ initialConfig: initialAgentConfig });
-  const [isToolSelectorOpen, setIsToolSelectorOpen] = useState(false);
-  const [isAddSubAgentModalOpen, setIsAddSubAgentModalOpen] = useState(false);
-  const [selectedAgentIdsForModal, setSelectedAgentIdsForModal] = useState<string[]>([]);
-
-  const handleAgentSelectionForModal = (agentId: string) => {
-    setSelectedAgentIdsForModal(prevSelectedIds => {
-      if (prevSelectedIds.includes(agentId)) {
-        return prevSelectedIds.filter(id => id !== agentId);
-      } else {
-        return [...prevSelectedIds, agentId];
-      }
-    });
-  };
-
-  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
-    const { name, value } = e.target;
-    updateField(name, value);
-  };
-
-  const handleLlmInputChange = (e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
-    if (config.type === AgentType.LLM) {
-      const { name, value } = e.target;
-      updateField(name, value);
-    }
-  };
-
-  const handleSliderChange = (value: number, field: keyof LlmAgentConfig) => {
-    if (config.type === AgentType.LLM) {
-      updateField(field, value);
-    }
-  };
-
-  const handleTypeChange = (newType: AgentType) => {
-    const newConfig = createNewAgentConfig(newType, config.id, config.name);
-    updateConfig(newConfig);
-  };
-
-  const handleSwitchChange = (checked: boolean, name: keyof LlmAgentConfig) => {
-    if (config.type === AgentType.LLM) {
-      updateField(name, checked);
-    }
-  };
-
-  const handleToolsSelectionChange = (selectedToolIds: string[]) => {
-    if (config.type === AgentType.LLM) {
-      updateField('tools', selectedToolIds);
-    }
-  };
-
-  const handleRemoveTool = (toolIdToRemove: string) => {
-    if (config.type === AgentType.LLM && config.tools) {
-      const updatedTools = config.tools.filter((toolId: string) => toolId !== toolIdToRemove);
-      updateField('tools', updatedTools);
-    }
-  };
-
-  const handleAddSubAgentToWorkflowByIds = (agentIdsToAdd: string[]) => {
-    const agentsToAdd: AnyAgentConfig[] = agentIdsToAdd
-      .map((id: string) => localMockExistingAgents.find((agent: AnyAgentConfig) => agent.id === id))
-      .filter(Boolean) as AnyAgentConfig[];
-
-    if (config.type === AgentType.Sequential) {
-      const currentAgents = config.agents || [];
-      const newAgentIds = new Set(currentAgents.map(a => a.id));
-      const uniqueNewAgents = agentsToAdd.filter(agent => !newAgentIds.has(agent.id));
-      updateField('agents', [...currentAgents, ...uniqueNewAgents]);
-    } else if (config.type === AgentType.Parallel) {
-      const currentAgents = config.agents || []; // parallel_agents is named 'agents' in ParallelAgentConfig type
-      const newAgentIds = new Set(currentAgents.map(a => a.id));
-      const uniqueNewAgents = agentsToAdd.filter(agent => !newAgentIds.has(agent.id));
-      updateField('agents', [...currentAgents, ...uniqueNewAgents]);
-    }
-    setIsAddSubAgentModalOpen(false);
-    setSelectedAgentIdsForModal([]);
-  };
-
-  const handleRemoveSubAgentFromWorkflow = (agentIdToRemove: string) => {
-    if (config.type === AgentType.Sequential) {
-      const currentAgents = config.agents || [];
-      const updatedAgents = currentAgents.filter((agent: AnyAgentConfig) => agent.id !== agentIdToRemove);
-      updateField('agents', updatedAgents);
-    } else if (config.type === AgentType.Parallel) {
-      const currentAgents = config.agents || [];
-      const updatedAgents = currentAgents.filter((agent: AnyAgentConfig) => agent.id !== agentIdToRemove);
-      updateField('agents', updatedAgents);
-    }
-  };
-
-  const handleSubAgentsOrderChange = (orderedSubAgents: AnyAgentConfig[]) => {
-    if (config.type === AgentType.Sequential || config.type === AgentType.Parallel) {
-      updateField('agents', orderedSubAgents);
-    }
-  };
-  
-  const handleLoopAgentChange = (selectedAgentId: string) => {
-    if (config.type === AgentType.Loop) {
-      const selectedAgent = localMockExistingAgents.find((agent: AnyAgentConfig) => agent.id === selectedAgentId);
-      if (selectedAgent) {
-        updateField('agent', selectedAgent);
-      }
-    }
-  };
-
-  const handleSavePress = async () => {
-    if (onSave) {
-      await onSave(config);
-    }
-  };
-
-  const llmConfig = config.type === AgentType.LLM ? config : null;
-  const sequentialConfig = config.type === AgentType.Sequential ? config : null;
-  const parallelConfig = config.type === AgentType.Parallel ? config : null;
-  const loopConfig = config.type === AgentType.Loop ? config : null;
-
-  const isSaveDisabled = 
-    !config.name ||
-    isSaving ||
-    (llmConfig && !llmConfig.instruction) ||
-    (sequentialConfig && (sequentialConfig.agents?.length || 0) < 1) ||
-    (parallelConfig && (parallelConfig.agents?.length || 0) < 1) || // Using 'agents' based on ParallelAgentConfig type
-    (loopConfig && !loopConfig.agent); // Check for the agent object itself
+    // updateConfig, // No longer directly used, type changes via handleTypeChange from hook
+    // updateField, // No longer directly used, input changes via handleInputChange from hook
+    llmConfig,
+    sequentialConfig,
+    parallelConfig,
+    loopConfig,
+    isToolSelectorOpen,
+    setIsToolSelectorOpen,
+    isAddSubAgentModalOpen,
+    setIsAddSubAgentModalOpen,
+    selectedAgentIdsForModal,
+    // setSelectedAgentIdsForModal, // Used by handleAgentSelectionForModal from hook
+    handleAgentSelectionForModal,
+    handleInputChange,
+    handleLlmInputChange,
+    handleSliderChange,
+    handleTypeChange,
+    handleSwitchChange,
+    handleToolsSelectionChange,
+    handleRemoveSelectedTool, // Renamed in hook
+    handleAddSubAgentToWorkflowByIds,
+    handleRemoveSubAgentFromWorkflow,
+    handleSubAgentsOrderChange,
+    handleLoopAgentChange,
+    handleSavePress,
+    isSaveDisabled,
+    localMockExistingAgents: hookLocalMockExistingAgents, // Use the one from the hook if needed, or keep component's
+  } = useAgentConfiguratorLogic({
+    initialAgentConfig,
+    onSave, // Pass onSave to the hook
+    isSavingGlobal: बाहेरून_येणारे_isSaving, // Pass isSaving to the hook
+  });
+  // MOCK_AVAILABLE_TOOLS is still needed for ToolSelector and rendering tool names.
+  // localMockExistingAgents from the component scope is used for AgentList filters.
+  // The hook also returns localMockExistingAgents, decide which one to use or if they should be synced.
+  // For now, assuming component's localMockExistingAgents for AgentList filters is fine.
 
   return (
     <div style={{ padding: '20px', maxWidth: '800px', margin: '0 auto' }}>
@@ -449,7 +316,7 @@ const AgentConfigurator: React.FC<AgentConfiguratorProps> = ({
                   return tool ? (
                     <Badge key={toolId} variant="secondary" className="flex items-center">
                       {tool.name}
-                      <Button variant="ghost" size="sm" className="ml-2 h-4 w-4 p-0" onClick={() => handleRemoveTool(toolId)}>
+                      <Button variant="ghost" size="sm" className="ml-2 h-4 w-4 p-0" onClick={() => handleRemoveSelectedTool(toolId)}>
                         <XIcon className="h-3 w-3" />
                       </Button>
                     </Badge>
@@ -584,10 +451,10 @@ const AgentConfigurator: React.FC<AgentConfiguratorProps> = ({
       <div style={{ marginTop: '30px', borderTop: '1px solid #eee', paddingTop: '20px' }}>
         <Button
           onClick={handleSavePress}
-          disabled={isSaveDisabled || isSaving}
+          disabled={isSaveDisabled} {/* isSaveDisabled from hook already considers isSavingGlobal */}
           style={{ minWidth: '120px' }}
         >
-          {isSaving ? 'Salvando...' : (isCreatingNew ? 'Criar Agente' : 'Salvar Alterações')}
+          {बाहेरून_येणारे_isSaving ? 'Salvando...' : (isCreatingNew ? 'Criar Agente' : 'Salvar Alterações')}
         </Button>
       </div>
     </div>
