+++ client/src/types/core/index.ts
@@ -1,0 +1,3 @@
+// client/src/types/core/index.ts
+export * from './agent';
+export * from './tool';
