+++ client/src/types/index.ts
@@ -1,4 +1,5 @@
-export * from './agent';
+export * from './core'; // Added
 export * from './chat';
-export * from './tool';
+// export * from './agent'; // Removed
+// export * from './tool'; // Removed
 export * from './adk';
